make the comments component stateless
The changes submitted in this commit includes creating a new
ConstComments component which is a stateless version of the
Comments component. The philosophy here is to use stateless
functional component as much as possible.
One improvement may be to cut the comments form out in its own
stateless functional component.